+++ src/main/groovy/com/cagnosolutions/starter/app/room/Room.groovy
@@ -21,7 +21,7 @@ class Room {
 	@GeneratedValue
 	Long id
 	String name
-	Float total
+	Double total
 	@OneToMany(cascade = CascadeType.ALL)
 	@JoinColumn(name = "room_id")
 	List<Item> items
@@ -30,8 +30,8 @@ class Room {
 		items << item
 	}
 
-	Float updateTotals() {
-		this.total = items*.updateTotal().sum() as Float
+	Double updateTotals() {
+		this.total = items*.updateTotal().sum() as Double
 	    this.total
     }
 }
